TTScore now 32bit, like Score

+++ src/zeta.cl
@@ -37,7 +37,7 @@ typedef char     s8;
 typedef u64     TTMove;
 typedef s32     MoveScore;
 typedef s16     Score;
-typedef s16     TTScore;
+typedef s32     TTScore;
 typedef u8      Square;
 typedef u8      Piece;
